Fix dose in egs++ 3ddose output
Fix the value of dose in 3ddose files generated from egs++ apps using
egs_dose_scoring. The dose was not being scaled by the mass of
materials, because the getMass function in the ND geometry was not
defined properly for XYZ geometries. This is a critical bug in the value
of dose reported and has been an issue since the ability to generate
3ddose files was added to egs_dose_scoring. In the limited case I
tested, the density used was always 1, instead of changing by material.